+++ src/main/java/org/lessons/java/shop/Smartphone.java
@@ -7,7 +7,7 @@ public class Smartphone extends Prodotto{
 	protected int codiceIMEI;
 	protected int memoria;
 	
-	public Smartphone(int codice, String nome, String marca, float prezzo, float iva,int codiceIMEI, int memoria) {
+	public Smartphone(int codice, String nome, String marca, int prezzo, int iva,int codiceIMEI, int memoria) {
 		super(codice, nome, marca, prezzo, iva);
 		this.codiceIMEI = codiceIMEI;
 		this.memoria = memoria;
