+++ src/main/java/org/lessons/java/shop/Televisori.java
@@ -5,7 +5,7 @@ public class Televisori extends Prodotto{
 	protected int dimensioni;
 	protected boolean smart;
 	
-	public Televisori(int codice, String nome, String marca, float prezzo, float iva, int dimensioni, boolean smart) {
+	public Televisori(int codice, String nome, String marca, int prezzo, float iva, int dimensioni, boolean smart) {
 		super(codice, nome, marca, prezzo, iva);
 		this.dimensioni = dimensioni;
 		this.smart = smart;
